+++ apps/editor.planx.uk/src/pages/Team/components/Archive.tsx
@@ -10,9 +10,9 @@ import ErrorSummary from "ui/shared/ErrorSummary/ErrorSummary";
 import { FlowCardView } from "../../FlowEditor/lib/store/editor";
 import { FlowTable } from "../components/FlowTable";
 import { ShowingServicesHeader } from "../components/ShowingServicesHeader";
-import { useGetArchivedFlows } from "../helpers/useGetArchivedFlows";
 import { DashboardList } from "./DashboardList";
 import FlowCard from "./FlowCard";
+import { useGetArchivedFlows } from "./hooks/useGetArchivedFlows";
 import { StyledToggleButton } from "./StyledToggleButton";
 type Props = {
   flowCardView: FlowCardView;
